牌阵更新

+++ src/services/cardarray.ts
@@ -1,4 +1,10 @@
-import { ArrayOption, GetOption, ModifyOption, Option } from "../types";
+import {
+  ArrayInterface,
+  ArrayOption,
+  GetOption,
+  ModifyOption,
+  Option,
+} from "../types";
 import CardArrayModel from "../model/cardArray";
 import { baseUrl } from "../config/config";
 import path from "path";
@@ -6,20 +12,50 @@ import fs from "fs";
 
 export class CardArray {
   async getarray(option: Option & GetOption) {
-    let data: any = [];
+    let data: Array<CardArrayModel> = [];
     //查询数据库
     try {
-      data = await CardArrayModel.findAll({
-        where:{
-          module:option.module,
+      if(option.status == 1){
+        data = await CardArrayModel.findAll({
+          where: {
+            module: option.module,
+            inuse: true,
+          },
+        });
+      }else if(option.status == 0){
+        data = await CardArrayModel.findAll({
+          where: {
+            module: option.module,
+            inuse: false,
+          },
+        });
+      }else{
+        data = await CardArrayModel.findAll({
+          where: {
+            module: option.module,
+          },
+        });
+      }
+      let sort: Array<CardArrayModel> = new Array<CardArrayModel>();
+      for (let i = 0; i < data.length; i++) {
+        data[i].src = baseUrl + `/cardArray/${data[i].src}`;
+        if (i == 0) {
+          sort.push(data[i]);
+        } else {
+          let len = sort.length;
+          for (let j = 0; j < len; j++) {
+            if (sort[j].pos > data[i].pos) {
+              sort.splice(j, 0, data[i]);
+              break;
+            } else if (j == len - 1) {
+              sort.push(data[i]);
+            }
+          }
         }
-      });
-      data.forEach((item: CardArrayModel) => {
-        //添加服务器访问地址
-        item.src = baseUrl + `/cardArray/${item.src}`;
-      });
+      }
+      data = sort;
     } catch (error) {
-      console.log(data);
+      console.error(data);
     } finally {
       return data;
     }
@@ -29,9 +65,27 @@ export class CardArray {
     return (await CardArrayModel.findAll()).length;
   }
 
-  async add(option: ArrayOption) {
-    //查询数据库
+  async sortAgain(index: number, module: string) {
+    let arrays = await CardArrayModel.findAll({
+      where: {
+        module,
+      },
+    });
+    arrays.forEach(async (item: CardArrayModel) => {
+      //检查是否有同名
+      if (item.pos < index) {
+        return;
+      } else {
+        item.pos++;
+        await item.save();
+      }
+    });
+    return true;
+  }
+
+  async add(option: ArrayOption & ArrayInterface) {
     try {
+      this.sortAgain(option.pos, option.module);
       await CardArrayModel.create({
         ...option,
       });
@@ -40,11 +94,11 @@ export class CardArray {
       return null;
     }
     let array = await CardArrayModel.findOne({
-        where:{
-            id:option.id
-        }
+      where: {
+        id: option.id,
+      },
     });
-   let data = array?.dataValues;
+    let data = array?.dataValues;
     return data;
   }
 
@@ -75,32 +129,36 @@ export class CardArray {
         },
       });
       if (array) {
-        if(array.src!= ''){
-            let filepath = path.join(__dirname, "../public/cardArray/", array.src);
-            fs.unlink(filepath,()=>{});
+        if (array.src != "") {
+          let filepath = path.join(
+            __dirname,
+            "../public/cardArray/",
+            array.src
+          );
+          fs.unlink(filepath, () => {});
         }
         await array.destroy();
       }
     } catch (error) {
-        console.error(error);
+      console.error(error);
       return null;
     }
     return "success";
   }
 
-  async delImage(id: string){
+  async delImage(id: string) {
     try {
       let array = await CardArrayModel.findOne({
         where: {
           id: id,
         },
       });
-      if (array && array.src != '') {
+      if (array && array.src != "") {
         let filepath = path.join(__dirname, "../public/cardArray/", array.src);
-        fs.unlink(filepath,()=>{});
+        fs.unlink(filepath, () => {});
         array.src = "";
         await array.save();
-      }else{
+      } else {
         return null;
       }
     } catch (error) {
